backend/requirement.txt
# Flask Core
Flask==3.0.0
flask-cors==4.0.0pyth
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
Flask-SQLAlchemy==3.1.1

# Security
werkzeug==3.0.1
PyJWT==2.8.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Utilities
user-agents==2.2.0
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-flask==1.3.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0